Insert placeholders for numeric properties when lowering object literals
Summary:
Simplify the logic for generating a serialized literal by always
serializing the key for non-serializable values, even when it is a
literal number. This will have a small size cost, in exchange for
creating the entire hidden class upfront, such that the subsequent
store goes to an existing property.

This is largely to avoid a problem in the IRVerifier, where the pass as
currently written can generate an empty `HBCAllocObjectFromBufferInst`,
which fails verification. Of course, other solutions are possible, and
perhaps more principled:

1. Allow empty `HBCAllocObjectFromBufferInst`, and pick a different
instruction at ISel.
2. Fix the pass here so that it calculates the number of properties
that will actually be serialized upfront.

However, since the serialized literal lowering is likely to be changed
soon anyway, this seems like the simplest solution that maintains the
current structure and invariants.

Reviewed By: avp

Differential Revision: D46787691

fbshipit-source-id: 3dce00e7bb1db5cdac3e68b4bb88c76b2329b8ba