Eliminate NewTarget when it is unused
Summary:
Add a check in `InstSimplify` to eliminate the `new.target` operand
when we know it will not be accessed. For now, this doesn't really make
a difference (or might even be worse since it requires loading
`undefined` in addition to the closure), but moving forward, it should
make it easier to do things like eliminate the closure or use a version
of call specialised for the case where `new.target` is `undefined`.

Reviewed By: tmikov

Differential Revision: D45195685

fbshipit-source-id: ddbce06275893e385edd756b2d28252222bd4f30