Handle dead Phis in simplifyPhiInst
Summary:
Original Author: neildhar@meta.com
Original Git: 40a28f0d00e204dd7155f5483a55785034abd429

It is possible to have a Phi in a dead block of the form:
```
%15 = PhiInst %15 : undefined|number, %BB4
```

This is a Phi with a single entry, which means that the current logic
will attempt to replace all uses of the Phi with itself, and then
delete it. The end result is that all existing users of the Phi
continue to hold the now freed Phi as an operand.

This can be fixed by simply deleting the special case for single entry
Phis, since the subsequent loop is more general and more correct. It
only replaces Phis where all entries refer to the same (non-self)
value, and performs no optimisation in the case where all entries are
self-edges.

Original Reviewed By: avp
Original Revision: D45760271

Reviewed By: neildhar

Differential Revision: D47690423

fbshipit-source-id: 5c74ec60e654ab9d1f077b323943d20a407f95d1