Move simplifySwitchInst to SimplifyCFG
Summary:
Original Author: avp@meta.com
Original Git: 709ad3d373430bd5f9c8325c2f99bba42942f07c

`simplifySwitchInst` modifies the CFG and can create unreachable basic
blocks, which need to be handled properly in SimplifyCFG.

The attached test was creating IR like this in the middle of
InstSimplify (abridged):
```
%BB3:
  %15 = PhiInst %19 : undefined|number, %BB4
%BB4:
  %19 = PhiInst %15 : undefined|number, %BB3
```
and running `simplifyPhiInst` on that would result in
```
%BB3:
  %15 = PhiInst %15 : undefined|number, %BB4
```
because for single-operand Phi we replace uses of the Phi with the
operand of the Phi, resulting in invalid IR.
This happens in unreachable blocks because the Phis can simply dominate
each other.

Moving `simplifySwitchInst` to SimplifyCFG fixes this by ensuring that
InstSimplify itself does not create unreachable blocks.

Next step is to only iterate the reachable blocks in InstSimplify.

Original Reviewed By: neildhar
Original Revision: D45738345

Reviewed By: neildhar

Differential Revision: D47690406

fbshipit-source-id: e42626b6a5782f9b5c4c778f70ae2aebdf83900e