Add asserts to limit large segments
Summary:
Original Author: neildhar@meta.com
Original Git: a4f16ec5c5f8156dc681be28daec3ca592a05997

Add an assert in the object creation path to more quickly detect if
an object is created with a size larger than the max.

Add a static assert in Hades when adding a new segment to the freelist
to verify that any region space in a segment can always be represented
in the size of a cell.

Original Reviewed By: tmikov

Original Revision: D46416780

Reviewed By: neildhar

Differential Revision: D47690539

fbshipit-source-id: 4ee5314a70cc69bde61a38ad5919c24025a98a27