Update Intl API documentation to include iOS support
Summary:
Original Author: jobpaardekooper@gmail.com
Original Git: beed7a01197237c6cf5463846f2d9a434c9429e2

As discussed in issue https://github.com/facebook/hermes/issues/1035,
the documentation of the Intl APIs does not include the current state
of the iOS support. This PR adds this information and formats the page
to reflect that this page now talks about Android and iOS support and
not just Android. Also the "Not yet supported" sections has been
removed. The page now only mentions what **is** supported and the
limitations of the functionalities that are currently supported. This
makes more sense with regard to the still changing specification.
Mentioning what is supported automatically means the remaining
information from the spec that is not included under "supported" is not
currently supported.

Original Reviewed By: avp

Original Revision: D47155521

Reviewed By: neildhar

Differential Revision: D47690636

fbshipit-source-id: 4829c12b2ff4fdfda4a5eeb07d3cb569154799c0