Fix null pointer to proxy handler
Summary:
Original Author: fbmal7@meta.com
Original Git: a43eb7bc44d8e340619b376641d4f62bcf52f9d9

Take a reference to the `handler` of the proxy before calling
`findTrap`, as we currently do for `target`. This respects the spec,
[10.5.8 [[Get]]](https://tc39.es/ecma262/#sec-proxy-object-internal-methods-and-internal-slots-get-p-receiver). According to spec, here is the ordering of events
```
1. Perform ? ValidateNonRevokedProxy(O).
2. Let target be O.[[ProxyTarget]].
3. Let handler be O.[[ProxyHandler]].
....
7. Let trapResult be ? Call(trap, handler, « target, P, Receiver »).
```

So, the very first thing we must do is take a reference to `handler`
once we have established we are not a revoked proxy. Then, this handler
is used as a reference elsewhere in the code.

Original Reviewed By: neildhar

Original Revision: D47347863

Reviewed By: neildhar

Differential Revision: D47690776

fbshipit-source-id: cb2435f66e870ecfa4cc4dc2deb2a3576616f098