Force all types to print in FlowTypesDumper
Summary:
We want to print every type in the dumper, but we could have a situation
in which we visit a type via another type before printing it.

e.g.
```
class %t.1 = class C {
  %constructor: function %t.2
  %homeObject: class %t.3
  x: number
}
class constructor %t.4 = class constructor class %t.1
```
In this case, we want to print `function %t.2` after printing the class,
but we wouldn't because the counter would've already progressed up to
4, and the `printAll` function would skip it.
The number has been allocated, but not via the main printing path.
Do the uniquing using a DenseSet instead, to account for ordering
problems like this one.

Reviewed By: tmikov

Differential Revision: D47959037

fbshipit-source-id: 7872f4451a4b680fe0ce51ef06acbd13e0b840e8