Fix resolution of BlockStatementNode
Summary:
BlockStatementNode was looking at the current node instead of parent
to know which nodes to skip.
This was dead code and now it's correctly checking the parent.

This error was mostly harmless for functions, because the extra scope
didn't actually get any declarations placed in it,
as the DeclCollector associated the declarations with the function.

Next step is to fix the CatchClause resolution because it does actually
need scopes associated with AST nodes, unlike functions.

Reviewed By: neildhar

Differential Revision: D47037407

fbshipit-source-id: 9175f8135738e8cb263fbe04d671ab4d0b1371b5