Fix resolution of variables in catch
Summary:
`catch` scopes weren't being handled correctly in the DeclCollector and
SemanticResolver. Variables are supposed to be associated with the catch
block because they need to share a scope with the catch params.

Putting the catch clause info in the DeclCollector also means that we
can handle it directly in `extractIdentsFromDecl` instead of specially
in the `CatchClauseNode` visitor.

This caused problems in the typechecker due to the mismatch in where
various parts of the pipeline were placing/looking up these variables.
Now, the typechecker can declare and typecheck that the catch param
ident is `any` or `mixed`, as is required.

Reviewed By: tmikov

Differential Revision: D47037406

fbshipit-source-id: 0345d315521efebc7816282117a32f4a4dabcc74