Make UnreachableInst into Terminator
Summary:
UnreachableInst should be usable as a terminator.
It's unreachable, there should be no code in the BasicBlock after it.

This allows us to use it in IRGen in the future in known-unreachable
locations even if we've emitted other code to get to that point,
such as for typechecked implicit returns we know can't ever run.

It was currently only emitted in one place where it was immmediately
followed by a Return anyway, so fix that spot.

Reviewed By: tmikov

Differential Revision: D48208213

fbshipit-source-id: 36d066d73e3c169993da00543098693e7be81ffc