MiniReact: Add Hello World element runner
Summary:
Use a new `createElement` to run some mini JSX through the render
1000000 times.

The benchmark can be made more elaborate in the future,
but this was the simplest way to write it to demonstrate it for now.

Also add a stripped version that can be run in engines without types.

Reviewed By: tmikov

Differential Revision: D47774552

fbshipit-source-id: 52bfa98b8c2648632b49a5f2f55166e315b6c173