Avoid revisiting var init expressions
Summary:
Variable init expressions were visited twice because the type associated
with the expression was used as a flag to avoid revisiting. Sometimes
that type would stay null, even though the init expression was visited.

This diff is a hack that uses a set to explicitly keep track of the
visited "inits". Ideally it may not be necessary and we may remove it
in the future because?
- Why is the type null? Shouldn't all expressions have types?
- Shouldn't we always visit the init expressions and skip here?

Reviewed By: avp

Differential Revision: D48296665

fbshipit-source-id: 077711574a5ac4e7f7aa32fb465d2963a7d33f42