Split FunctionType into typed/untyped
Summary:
Allow FunctionType to indicate whether the
signature contains annotations at all.
If a single annotation exists, then the function is "typed",
and is represented with `TypedFunctionType`.
Otherwise it's "untyped" (i.e. calls should be treated as legacy)
and we use `UntypedFunctionType`.

Populate the `typed` attribute in the IR based on that.

Update some tests to account for `main` functions being untyped.

Reviewed By: tmikov

Differential Revision: D46806650

fbshipit-source-id: b3ad2765466ebc08d9255fc3fcac2e7f59e989ae