Remove nullptr 'any' parse in parseTypeAnnotation
Summary:
`parseTypeAnnotation` should only be used if we actually have something
to parse, and if there's no type annotation that should be handled
explicitly at the callsite to avoid unexpectedly receiving `any`.

This removes an instance of a random `any` by replacing the check with
an assert.

Audited the callsites and none of them were potentially calling with
`nullptr` anyway.

Reviewed By: tmikov

Differential Revision: D46772355

fbshipit-source-id: 384b2676df8cbdb4943abb5b59996b5d181bbb74