Emit UnreachableInst for unreachable implicit returns
Summary:
We have information in FunctionInfo to let us know whether a function
will have a reachable implicit return.
Use that to avoid emitting a potentially incorrectly typed `undefined`
return.
Do this change in both typed and untyped code to allow for better
testing.

The only non-test changes are in `lib/IRGen`.

Reviewed By: tmikov

Differential Revision: D48249308

fbshipit-source-id: a98e2e4fe62d5106d3431d005bb0f8578e7905a2