Add NativeContext
Summary:
`NativeContext` is an extension of `Context`, deliberately kept separate
for simplicity. It owns native types, native function signatures and
extern native function declarations.

The target machine is configured by a LLVM triple, parsed by whatever
we have in LLVH, which is probably outdated, but sufficient for our
purposes for now.

The target is specified using the hidden (since it is not user
friendly) flag `-Xnative-target`. It is not validated. Any triple,
including random garbage, can be supplied. However it does work when
valid 32-bit and 64-bit triples are passed, and was tested higher in
the stack.

NativeContext is not used yet.

Reviewed By: avp

Differential Revision: D48720345

fbshipit-source-id: 203d4ae2d123bcce6ba5f8bde6036195143c0b5f