ffigen: a simple tool to generate FFI bindings from C headers
Summary:
The tool is very primitive, but it does successfully generate all
bindings used by the sqlite.js example.

It still doesn't recognize all types. In particular it needs to be
extended to follow through typedefs, but improvements like that can be
endlessly iterated on.

The tool uses `castxml`, which needs to be installed. It is included in
Homebrew for MacOS.

Reviewed By: neildhar

Differential Revision: D48938801

fbshipit-source-id: 12836a225ef37a208841a863cf2d712f09d1883b