ffigen: EASY: skip some comments when grepping for defines
Summary:
The previous regex would not detect defines with the following format:
```
#define MYCONST 100 /* my const */
```
It was looking for numbers followed by spaces and new line. The new
regex is basically a hack that looks for numbers followed by spaces and
a new line or start of comment.

Clearly, it is not always correct, like in this (hypothetical) case:
```
#define BAD 100 /* evil comment */ 200
```

However this is just best effort and works well enough for sqlite3.h.

Eventually we should do proper parsing, for which a regex is not the
best tool. Since parsing in Python would be exceptionally slow, we will
probably still use a "loose" regex to find all candidate lines starting
with "^\s*#\s*define\s+\w+" and then manually parse them.

Reviewed By: avp

Differential Revision: D49027842

fbshipit-source-id: ae0420887a95490f4613cda893d10e05e70a3d01