Ignore functions with no callsites in FuncSigOpts
Summary:
If we have a function without callsites, we can't replace the arguments
within the function because they could, e.g. require certain types
that FuncSigOpts doesn't concern itself with.

Take this purely constructed example:
```
function foo(x: object): any [allCallsitesKnownInStrictMode,typed]
  %1 = LoadParamInst (:object) %x: object
  %2 = PrLoadInst (:number) %1: %object, 0: number, "y": string
```

If we were to replace `x` when `foo` had no callsites,
we'd violate the requirement that the first operand of `PrLoad` is
an `object`.

This would cause problems more concretely in nbody,
where adding the FUnaryMath and FBinaryMath instructions means
that we have more instructions with enforced constraints on the types of
their operands.

NOTE: In most cases, function should be deleted in the future anyway.
However, functions called with `new` are retained so we can keep the
`.prototype` property.
This means that constructor functions which are inlined at all
callsites will not participate in FuncSigOpts.
A potential future solution may be to delete the bodies of the functions
knowing that they are never called, allowing us to delete all the
parameters as well.

Reviewed By: neildhar

Differential Revision: D49008234

fbshipit-source-id: 3eb5859eb3c8881c380e377468989c5d88f664a5