Run InstSimplify iteratively by inserting after instructions
Summary:
If we simplify an instruction, it's possible that we could simplify it
further by running `InstSimplify`'s checks on it again.
We can get this easily by inserting after the optimized instruction
and letting the loop pick up the new instruction on the next iteration.

Make it hard to make a mistake by centralizing the `builder_` in an
Impl class.
While we're here, remove the unnecessary header file.

Reviewed By: tmikov

Differential Revision: D49158317

fbshipit-source-id: 8f73c7a9fde65e0c0fe108323315134a3405e604