Print operand deps in allocated IR
Summary:
Since we decided that we are going to preserve SSA after register
allocation, it now makes sense to print the operand dependencies again,
in addition to allocated registers.

For simplicity we always print everything, there are no command line
options to turn one or the other off. It has become a little busy, so
we had to experiment a bit to make both the allocated register and the
operand visible.

The next diff arguably further improves it a little bit.

While we are here we also improved the printing of register liveness
intervals when enabled. Slightly better alignment and using a consistent
syntax for instruction numbers.

Reviewed By: avp

Differential Revision: D49357877

fbshipit-source-id: 918a494b58a3d1c3ebfe347e9af8ea8c9b97c31c