Add implicit checked cast to AssignmentExpression
Summary:
To allow `a += b` when `a` is typed (e.g. `number`) but `b` is `any`,
FlowChecker has to insert ImplicitCheckedCast around the result of `+`,
but there's no node that exists for that.

Handle the problem by putting the ImplicitCheckedCast on the LHS,
and dealing with this case specially in IRGen by putting the cast
on the result of the binary operation.

This results in `interp-dispatch` being able to run with `-typed`
without any annotations, because `res += fact` works now.
`res` is inferred to be `number` based on its initializer,
but `fact` is `any`.

Reviewed By: tmikov

Differential Revision: D49471684

fbshipit-source-id: 63b2cd313fb02c0fc375fe0fc500320c1db6d04b