Strengthen invariant for SNOP list in LowerAllocObject
Summary:
Original Author: neildhar@meta.com
Original Git: 1ec38ec48c9153e8c62e6fd5afd2a7247316dbbd

We should stop processing `StoreNewOwnPropertyInst`s as soon as we find
a usage of the `allocObject_` where it is not the target of a SNOP. To
keep the invariant simple this should include the case where an object
is stored as a property on itself.

I don't think it's actually possible to emit this IR currently, but
this clarifies the invariant.

This invariant is also convenient for the next diff, since it means
that each serialisable SNOP may only account for a single use of an
`AllocObjectInst`.

Original Reviewed By: avp

Original Revision: D30649306

Reviewed By: avp

Differential Revision: D50460153

fbshipit-source-id: 7139a30d35c85e223e9c1d5dae7a82e1d9557273