ScopedHashTable: Fix insert semantics and naming
Summary:
There were multiple things wrong with ScopedHashTable::insert().

First, it overwrote the value if the key already existed in the scope,
even though its name implied that it wouldn't (similarly to standard
containers). The impression that it doesn't overwrite was strengthened
by the existence of the method `setInCurrentScope()`, which was
unnecessary.

Second, when overwriting an existing value, it would do it by inserting
a new key/value pair and shadowing the previous one using the
`nextShadowed_` link. That link is intended for shadowing values in
upper scopes, not in the same scope. Even though the old value was kept
in the table in this way, it could no longer be accessed in any way.

Except, `eraseFromCurrentScope()` would erase the current value and
restore the previous one, if it existed. That is not what the poor
misguided author of that method (it was me) intended.

Address all of this by:
1. Use `std::pair` to store the keys and values to match standard
semantics.
2. Renaming `insert()` to `try_emplace()`, since it takes its arguments
unpacked.
3. Change it to return an iterator/bool pair, matching standard
semantics.
4. Implement `setInCurrentScope()` (which now makes sense) in terms of
the now correct `try_emplace()`.

Reviewed By: neildhar

Differential Revision: D50764925

fbshipit-source-id: 66f7a973684b9f6cf519b6fe30d9541b268cbd14