Allow 'of' as first identifier of a yield argument
Summary:
Original Author: avp@meta.com
Original Git: 43fd8774825e77eed38570acd05a87e7d8811e44
Original Reviewed By: tmikov
Original Revision: D46728601

`yield of()` was erroring in the parser due to the
`checkEndAssignmentExpression` check returning `true` on `of`,
because of `for..of` loops.

But in a `yield` expression `of` doesn't terminate it,
so allow a way to exclude it.

Reviewed By: neildhar

Differential Revision: D50992838

fbshipit-source-id: fabd1a8a6640ffdce78576499f2f8c41cf3c9681