improve debug location caching with a source map
Summary:
Original Author: tmikov@meta.com
Original Git: 087fd7356073c68aa0a6d1ca05bcc7e082a4dee1
Original Reviewed By: avp
Original Revision: D48633068

In some extreme cases when a source map and a very large source mapping
URL are used, the debug location generation could slow down dramatically
(going from seconds to minutes).

Specifically the slowdown is in `HBCISel::getDebugSourceLocation()`, in
the part that records the new filename and source mapping URL. If the
filename, **or particularly the source mapping URL**, are very large,
the hashmap lookup can be very expensive, since it needs to scan the
entire string at three times (to calculate the hash and to compare).

The code used a simple caching scheme, with a single cache entry
recording the last buffer id to avoid the expensive operations above.
Unfortunately even with that level of caching, the operations can occur
tends of thousands of times, resulting in dramatic slowdowns.

Perhaps a better solution is to deduplicate these strings before this
point, but that would be a too disruptive change.

Instead, we improve the cache by replacing the single cache entry with a
hash map, mapping from buffer id to filename and source mapping URL. The
hash map can only grow up to the number of input source files in the source
map, so it is bounded.

Additionally, we save a pointer to the last found entry in the map, so
in the fast path we can even avoid the hash map lookup.

Previously the single cache entry was shared between functions by
manually getting and setting it in instances of `HBCISel`. We clean this
up by simply passing the "cache" as a reference to the `HBCISel`
constructor.

Reviewed By: tmikov, neildhar

Differential Revision: D50992854

fbshipit-source-id: 02d817a0e58398407dcd1b709673f80a3f6ff7fd