Convert 'as' to a new token
Summary:
Original Author: avp@meta.com
Original Git: b2796f7c71768c7467dc176f84b3ab6869a40854
Original Reviewed By: tmikov
Original Revision: D49700751

Fixes the parsing of multiple TypeScript as expressions in a row.
Previously `x as any as number` was parsed as `x as (any as number)`,
instead of `(x as any) as number`.

This was because we customized the precedence calculation for `as`
in the function getPrecedenceExcept,
but this custom precedence was not saved in the stack.
Later when we had an as expression in the stack,
we were not getting the right precedence level for it when doing
`getPrecedence(stack.back().opKind)`.

Fix this by checking for `identifier` specially and converting
the `as` identifier to a new token kind, `as_operator`, before pushing
onto the stack.
This can be extended to implement `satisfies` as well.

Reviewed By: tmikov

Differential Revision: D50992880

fbshipit-source-id: 299673a3f36f0ba486c728a98cdc5703cc96653c