Add a DCE pass after inlining
Summary:
Add an extra invocation of DCE after inlining to improve stack
promotion.

Stack promotion after inlining can't do its job if frame variables
appear to be accessed from the body of an inlined function, which is now
dead. Running DCE after inlining improves this by removing the dead
functions.

The following example illustrates the problem:
```
    (function () {
        var x = 10;
        function leaf() {
            x += x + 1;
        }
        function middle() {
            for(var i = 0; i < 5; ++i)
                leaf();
        }
        middle();
        return x;
    })();
```

After middle() and leaf() are inlined, the now dead body of leaf() is
not removed, forcing `x` to stay as a frame variable.

Note that the current implementation of DCE itself has limitations (see
T170320383) so the newly inserted DCE pass actually doesn't remove
`leaf()`. But it does a little additional work which makes it possible
to a later pass to actually achieve what this first pass should have.

With that said, a DCE pass after inlining and before stack promotion is
the logical place for this, so hopefully when DCE is improved, the
optimization will simply happen earlier.

Added a new test showing this. A couple of other tests improved a
little.

Reviewed By: neildhar

Differential Revision: D51418640

fbshipit-source-id: dce3535a15915268212865a85e294bf3d07ee643