Only call canBeInlined once per function
Summary:
`canBeInlined` was iterating the whole function on every callsite.
This wasn't necessary, as the only check in `canBeInlined` that cared
about the callsite was the recursion check.

Move the recursion check directly into the callsite loop,
and `canBeInlined` can now just be called from outside the loop
as a check at the start of `shouldTryToInline`.

Reviewed By: neildhar

Differential Revision: D51181084

fbshipit-source-id: fef86db3ef0a656cd68f8cfab75d2a23a7b5a079