Allow inlining small functions with multiple callsites
Summary:
Small functions can always be inlined, even if we don't know every
callsite or if it has multiple callsites.
This cuts down on call overhead and can even save code size if
optimization is able to constant fold, delete variables, etc.

For now, this is only enabled for functions that have numArgs+1
instructions (excluding LoadParam and Return, because those are always
going to disappear when inlining).

Future steps include being able to configure this from the CLI.

Reviewed By: neildhar

Differential Revision: D51163399

fbshipit-source-id: f1ca87a08ae61adb92804b92be98f87586ff67ee