InstSimplify: fix handling of undefined in relational comparisons
Summary:
Relational comparisons (<= and >=) are simplified to a true/false when
it is known that the operands are the same value with a primitive type.
If the value is "undefined", it is converted to NaN, which never equals
itself.
The optimization was not correctly handling the case where the operand
could be "undefined" or another type - it was only checking for
"undefined".

Example:

    try {} catch {
      var v0 = "m";
    }
    return v0 <= v0;

The try/catch is necessary to make the optimizer believe `v0` is not a
constant.

Reported in https://github.com/facebook/hermes/issues/1200.

Reviewed By: avp

Differential Revision: D51555317

fbshipit-source-id: 55bad0a8ce2a9ed233d0609b9480c1636fc6a658