Prohibit unreachable blocks in between passes
Summary:
Enforce a new invariant in the IRVerifier that there must not be
unreachable blocks between compiler passes. Add calls to delete
unreachable blocks to IRGen and inlining to maintain this invariant.

This tries to eliminate the large class of problems that come from
keeping unreachable code around. In particular, how it might continue
to maintain references to other reachable or unreachable code.

Reviewed By: avp

Differential Revision: D51241966

fbshipit-source-id: 52eea55378bbe5ee3a1221cb4bb2f5858f80f4cb