Fix DCE function deletion
Summary:
Dead function deletion can currently leave dangling references to the
deleted function if inner functions access the variables of the deleted
function but cannot themselves be deleted.

To address this, change the deletion function to also delete all users
of a function that is identified as being dead, since we know that they
must also be dead.

Reviewed By: avp

Differential Revision: D50600299

fbshipit-source-id: 62e07199f100b419c98282917470c92d3b203ca7