Always write result last in the interpreter (#1203)
Summary:
Original Author: neildhar@meta.com
Original Git: 1da8a1f94071ddaf42c844d6b71452ffe8f6e833

If an instruction is the last user of one of its operands, the register
allocator may reuse one of its operands as the result register. This is
typically fine since most operands are only input parameters, and they
are always read before the output is written.

However, when the operand is both an input and an output, this requires
care to ensure that the operand register (whose live range ends with
the given instruction) is written before the result. This ensures that
even if the register allocator assigns them the same register, the
value held in that register after the instruction is the result.

Original Reviewed By: avp

Original Revision: D51597086

Reviewed By: avp

Differential Revision: D51835093

fbshipit-source-id: e20baae1ff33bc8bad2565ccb0034e85a990b97b