shermes: add `--dump-transpiled-ast`
Summary:
I noticed that `--dump-ast` was outputting AST after TS-to-Flow
transpilation, and also after IIFE wrapping.

I found that to be confusing, since I expect "--dump-ast" to literally
print what was parsed, before any transformations have been performed on
it.

Move the dumping of the parsed AST earlier in the code and add a new
dumping stage `--dump-transpiled-ast`, which prints the AST after the
optional early transpilation to Flow and IIFE wrapping.

Reviewed By: avp

Differential Revision: D52149761

fbshipit-source-id: 56a83e9c42e0286faa164f0bdefe94883eaf11da