SourceMap improvements
Summary:
This diff contains two improvements:

## StringRef
SourceMap's API was defined to accept and return `std::string`, causing
many unnecessary allocations and copies. Specifically, there was a new
`std::string` for every translated location. Use `llvh::StringRef`
instead.

## BaseDir
Extend SourceMapParser and SourceMap with an optional path prefix. A
source map usually contains paths relative to itself. When accessing a
source map from a directory, the paths are relative to that directory.
In order to use those paths, that directory has to be prepended to each
input path.

Example:

`dir1/file.ts` is transpiled to `dir1/out/file.js` and
`dir1/out/file.js.map` by invoking `tsc` from within `dir1/`. The source
map `dir1/out/file.js.map` points to `file.ts` using a relative path
`../file.ts`.

If we were to open `/tmp/dir1/out/file.js.map`, we need to prepend
`/tmp/dir1/out/` to the relative path `../file.ts` in order to obtain
the correct path to the original source.

This could be done in a layer above SourceMap, but it is just easier and
simpler to do it inside SourceMap, since it already has support for
`sourceRoot`. Plus, when we are translating hundreds of thousands of
locations, we don't want to perform the prepending every time.

Reviewed By: ftanuma

Differential Revision: D52062316

fbshipit-source-id: 337e841ec90647f19c7a9b4c041dae99ac292745