Always avoid typechecking when declaring scope types fails
Summary:
`resolveScopeTypesAndAnnotate` can report errors and leave the
annotated AST in an invalid state when it fails.
We should not proceed with typechecking in this case.

We were already doing this in ProgramNode, but it has to be done for
every node that calls the function, so add `LLVM_NODISCARD` and return
a success flag.

Reviewed By: tmikov

Differential Revision: D51644092

fbshipit-source-id: 4fa9bc1ebe211066d07fbe42b679e3d030da7962