Complete forward declared unions before parsing class types
Summary:
If we parse classes prior to looking at the union declarations and
canonicalizing them, then we are referring back to the type aliases
before they are able to be flattened into more unions.

```
type A = number | string;
class C {
  x: A | null;
}
```

We need to canonicalize `A` and determine whether it's actually a union
before parsing the `A | null` inside `C`, otherwise we violate the
assumption that we have collapsed nested unions.

Reviewed By: tmikov

Differential Revision: D51644094

fbshipit-source-id: 39c5e3af06cd61f2af48fd33b456419a0a642546